requirements.txt
mlflow>=2.12.0
psycopg2-binary>=2.9.5
